Tokenizer/requirements.txt
azure-common==1.1.27
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.6
azure-functions==1.7.2
azure-nspkg==3.0.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
idna==3.3
pycparser==2.21
python-dateutil==2.8.2
requests==2.26.0
six==1.16.0
urllib3==1.26.7
